REFACTORED
