+++ inbus-front/public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-<!--    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />-->
+    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
@@ -11,12 +11,11 @@
     />
     <link rel="apple-touch-icon" href="logo192.png" />
 
-<!--    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />-->
+    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
     <title>React App</title>
   </head>
   <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
   </body>
 </html>
