+++ inbus-front/src/App.css
@@ -1,7 +1,3 @@
-/*.flexBox {*/
-/*  display: flex;*/
-/*}*/
-
 .App {
   display: flex;
 }
