+++ inbus-front/src/pages/aboutus/AboutUs.css
@@ -3,6 +3,7 @@ main {
     margin-left: 25px;
     background-color: #dad8c1;
     border-radius: 0 0 5px 5px;
+    text-align: center;
 }
 
 main header {
@@ -23,7 +24,6 @@ main h3 {
 main h4 {
     font-family: Roboto, sans-serif;
     font-size: 22px;
-    /*margin: 25px 0 25px 390px;*/
 }
 
 main p {
@@ -37,6 +37,6 @@ main footer p {
     font-size: 12px;
 }
 
-main footer .sponsored {
+main footer .sponsor {
     padding-bottom: 15px;
 }
