+++ inbus-front/src/pages/orders/components/orderdescription/OrderDescription.jsx
@@ -1,0 +1,16 @@
+import React, {Component} from "react";
+import './orderdescription.css'
+
+class OrderDescription extends Component {
+    render() {
+        return (
+            <div className="description">
+                <h4>Описание</h4>
+                <p>{this.props.description}</p>
+                <img src="./gps_label.png" alt="MAP"/>
+            </div>
+        );
+    }
+}
+
+export default OrderDescription
