+++ inbus-front/src/pages/orders/components/orderdescription/orderdescription.css
@@ -1,0 +1,16 @@
+.description {
+    text-align: center;
+}
+
+.description h4 {
+    font-family: Roboto, sans-serif;
+    font-size: 16px;
+    padding-top: 25px;
+    margin: 0 0 10px 0;
+}
+
+.description p {
+    font-family: Roboto, sans-serif;
+    font-size: 14px;
+    padding: 10px;
+}
